US10-DISTANCE DATA AND POINTS INTEGRATION:
- Altered reward_points field to FloatField for better precision.
- Updated cumulative distance handling in the user model.
- Modified reward account view to include cumulative distance.
- Enhanced RewardsPage to display cumulative distance and updated date dynamically.

+++ backend/app/rp_core/models.py
@@ -10,8 +10,8 @@ class AppUser(AbstractUser):
     # Override username to allow duplicates
     username = models.CharField(max_length=150, unique=False)
     email = models.EmailField(unique=True)
-    reward_points = models.PositiveIntegerField(
-        default=0,
+    reward_points = models.FloatField(
+        default=0.0,
         help_text="Points available for redeeming rewards.",
     )
     # Total cumulative distance the user has travelled using the app (in kilometres).
